# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Your Auth API",
    "description": "API de autenticacao e identidade para contas proprietarias, projetos, usuarios finais, sessoes, refresh tokens e API keys de projeto.",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/projects/{projectId}/auth-config": {
      "get": {
        "tags": [
          "Project Auth Config"
        ],
        "summary": "Busca a configuracao de autenticacao de um projeto",
        "operationId": "getAuthConfig",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuracao encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConfigDTO"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou configuracao nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "Project Auth Config"
        ],
        "summary": "Atualiza a configuracao de autenticacao de um projeto",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthConfigDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuracao atualizada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthConfigDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou configuracao nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "AuthConfigDTO": {
        "type": "object",
        "properties": {
          "accessTokenExpirationMinutes": {
            "type": "integer",
            "format": "int32",
            "maximum": 1440,
            "minimum": 1
          },
          "refreshTokenExpirationDays": {
            "type": "integer",
            "format": "int32",
            "maximum": 365,
            "minimum": 1
          },
          "sessionMode": {
            "type": "string",
            "enum": [
              "MULTIPLE_DEVICES",
              "SINGLE_ACTIVE_SESSION",
              "LIMITED_ACTIVE_SESSIONS"
            ]
          },
          "maxActiveSessions": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          },
          "refreshTokenRotationEnabled": {
            "type": "boolean"
          },
          "revokeTokensOnPasswordChange": {
            "type": "boolean"
          },
          "failedLoginAttemptsLimit": {
            "type": "integer",
            "format": "int32",
            "maximum": 20,
            "minimum": 1
          },
          "lockDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "maximum": 1440,
            "minimum": 1
          },
          "requireEmailVerification": {
            "type": "boolean"
          },
          "registrationEnabled": {
            "type": "boolean"
          },
          "maxActiveSessionsRequiredWhenLimited": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT de conta proprietaria enviado no header Authorization: Bearer <token>.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "projectApiKey": {
        "type": "apiKey",
        "description": "API key de projeto usada por clientes terceiros. Tambem pode ser enviada como Authorization: Bearer ya_sk_...",
        "name": "X-API-Key",
        "in": "header"
      },
      "accessTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para autenticacao.",
        "name": "access-token",
        "in": "cookie"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para renovar a sessao.",
        "name": "refresh_token",
        "in": "cookie"
      }
    }
  }
}